+++ example/midway-react-ssr/config.ts
@@ -1,7 +1,7 @@
 import type { UserConfig } from 'ssr-types'
 
 const userConfig: UserConfig = {
-  dynamic: false
+
 }
 
 export { userConfig }