+++ packages/server-utils/src/webpack/externals.ts
@@ -1,9 +1,11 @@
 // @ts-nocheck
+import { resolve } from 'path'
 import { contains, containsPattern, readFromPackageJson, readDir } from './external-utils'
+import { getCwd } from '../cwd'
 
 const scopedModuleRegex = new RegExp('@[a-zA-Z0-9][\\w-.]+\/[a-zA-Z0-9][\\w-.]+([a-zA-Z0-9.\/]+)?', 'g')
 
-function getModuleName (request, includeAbsolutePaths) {
+function getModuleName(request, includeAbsolutePaths) {
   let req = request
   const delimiter = '/'
 
@@ -19,16 +21,50 @@ function getModuleName (request, includeAbsolutePaths) {
   return req.split(delimiter)[0]
 }
 
-function nodeExternals (options) {
+function wrap(whitelist) {
+  const deepWhiteList = [...whitelist]
+  const map = {}
+  function getDep(whitelist, deepWhiteList) {
+    for (const dep of whitelist) {
+      if (typeof dep !== 'string') {
+        continue
+      }
+      try {
+        if (!map[dep]) {
+          deepWhiteList.push(new RegExp(dep))
+          map[dep] = true
+        }
+        const childWhiteList = []
+        const packageJSON = require(resolve(getCwd(), `./node_modules/${dep}/package.json`))
+        for (const childDep in packageJSON.dependencies) {
+          if (!map[childDep]) {
+            deepWhiteList.push(new RegExp(childDep))
+            map[childDep] = true
+          }
+          childWhiteList.push(childDep)
+          getDep(childWhiteList, deepWhiteList)
+        }
+      } catch (error) {
+
+      }
+    }
+  }
+  getDep(whitelist, deepWhiteList)
+  return deepWhiteList
+}
+
+
+function nodeExternals(options) {
   options = options || {}
-  const whitelist = [].concat(options.whitelist || [])
+  let whitelist = [].concat(options.whitelist || []);
+  whitelist = wrap(whitelist)
   const binaryDirs = [].concat(options.binaryDirs || ['.bin'])
   const importType = options.importType || 'commonjs'
   const modulesDir = options.modulesDir || 'node_modules'
   const modulesFromFile = !!options.modulesFromFile
   const includeAbsolutePaths = !!options.includeAbsolutePaths
   // helper function
-  function isNotBinary (x) {
+  function isNotBinary(x) {
     return !contains(binaryDirs, x)
   }