+++ packages/plugin-vue/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.7.15](https://github.com/zhangyuang/ssr/compare/ssr-plugin-vue@5.7.14...ssr-plugin-vue@5.7.15) (2021-12-02)
+
+**Note:** Version bump only for package ssr-plugin-vue
+
+
+
+
+
+## [5.7.14](https://github.com/zhangyuang/ssr/compare/ssr-plugin-vue@5.7.13...ssr-plugin-vue@5.7.14) (2021-12-02)
+
+**Note:** Version bump only for package ssr-plugin-vue
+
+
+
+
+
+## [5.7.13](https://github.com/zhangyuang/ssr/compare/ssr-plugin-vue@5.7.12...ssr-plugin-vue@5.7.13) (2021-12-02)
+
+**Note:** Version bump only for package ssr-plugin-vue
+
+
+
+
+
+## [5.7.12](https://github.com/zhangyuang/ssr/compare/ssr-plugin-vue@5.7.11...ssr-plugin-vue@5.7.12) (2021-12-02)
+
+**Note:** Version bump only for package ssr-plugin-vue
+
+
+
+
+
 ## [5.7.11](https://github.com/zhangyuang/ssr/compare/ssr-plugin-vue@5.7.10...ssr-plugin-vue@5.7.11) (2021-11-30)
 
 **Note:** Version bump only for package ssr-plugin-vue